finder/requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
PyMuPDF
python-docx

